+++ CHANGELOG.md
@@ -2,7 +2,9 @@
   
 - **[command_menu] feature: now we have an `online menu editor`, which greatly helps you design a menu. (Thanks to @Hitnam)**
   - You can use the editor in https://fuji-command-menu-editor-k4k4.vercel.app/
-- [motd] feature: optimize the `default config file` for this module. (Now it provide the fancy gradient colors in the default example)
+- [motd] feature: improvements to `motd` module.
+  - feature: now allow to customize the `players info` area. (You can customize the `max players`, `online players` and `hover text`.)
+  - feature: optimize the `default config file` for this module. (Now it provide the fancy gradient colors in the default example)
 - [tab] feature: optimize the `defualt config file` for this module. (Now it provide a better color schema in the default example)
 - [fuji] feature: now allows the `console` to execute `/fuji {reload/about/debug}` commands.
 - [core] feature: improvements to `language` system.